+++ NQueen.Benchmarking/Usings.cs
@@ -7,7 +7,6 @@ global using NQueen.Domain.Enums;
 global using NQueen.Domain.Interfaces;
 global using NQueen.Domain.Models;
 global using NQueen.Domain.Utils;
-global using NQueen.Kernel;
 global using NQueen.Kernel.Solvers;
 global using NQueen.Kernel.Solvers.Engines;
 global using System;